chore: move features/filterbuilder.vue to components/filterbuilder.vue
